Merged PR 95: 34595 - Add banned words list
Created a SharePoint list and that contains a list of banned words for the username.
Added TypeScript to pick up this list and return "Name is not available for use." if a banned word is detected in the first or last name. This also turns the "Username Validation" box to make it easily identifiable to the user that the name cannot be used.

Related work items: #34595